edit機能をnew機能へ変更 ストレージをモジュール化
