+++ src/views/Home.vue
@@ -6,7 +6,7 @@
         <li v-for="item in memos" v-bind:key="item.id">
           <router-link :to="{ path: '/memo/' + item.id }" v-on:click.prevent="showForm">{{ item.content.split('\n')[0] }}</router-link>
         </li>
-        <li><router-link to="/edit" v-on:click.prevent="showForm">+</router-link></li>
+        <li><router-link to="/new" v-on:click.prevent="showForm">+</router-link></li>
       </ul>
       <div class="form-box">
        
@@ -15,29 +15,9 @@
   </div>
 </template>
 <script>
-let STORAGE_KEY = 'memos-vuejs-demo'
-let memoStorage = {
-  fetch: function() {
-    let memos = JSON.parse(
-      localStorage.getItem(STORAGE_KEY) || '[]'
-    )
-    memos.forEach(function(todo, index) {
-      todo.id = index
-    })
-    memoStorage.uid = memos.length
-    return memos
-  },
-  save: function(memos) {
-    localStorage.setItem(STORAGE_KEY, JSON.stringify(memos))
-  }
-}
-
-// import Edit from './views/Edit.vue'
+import { memoStorage } from "../modules/storage";
 export default {
   name: 'App',
-  // components: {
-  //   Edit
-  // },
   data () {
     return {
       memos: [],
@@ -47,7 +27,6 @@ export default {
   methods: {
     showForm: function() {
       this.showText = true
-      console.log(this.memos);
     },
     addMemo: function() {
       let content = this.$refs.content
